pyproject.toml
[project]
name = "fastapi-ecommerce-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioredis>=2.0.1",
    "aiosmtplib>=5.1.0",
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi-cache2[redis]>=0.2.2",
    "fastapi[all]>=0.128.0",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
    "ty>=0.0.13",
]

[tool.ruff]
# Target Python version
target-version = "py314"

# Line Length (FastApi-friendly)
line-length = 100
indent-width = 4


exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".ruff_cache",
    "alembic/versions"
]

# Enable autofix by default (ruff --fix)
fix = true

[tool.ruff.lint.isort]

# Import sorting styles
combine-as-imports = true
force-single-line = false
known-first-party = ["app", "core"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"

indent-style = "space"

line-ending = "auto"
